Fix newlines in log files
If we set the encoding of the file itself to 'unicode-escape' then we
get no unicode errors, but we get no newlines either. The \n in the
string gets escaped I suppose.

This patch sets the encoding to binary, then encodes each line of text
we write using unicode-escape explicitly. This allows writing \n, and
produces more useful results although who knows there may be a more
correct way to solve this.