Continue sequence on hosts that haven't failed, remove --ignore-errors
In practice I rarely want my nightly builds to abort because of an
issue on one host. Instead, I want as many tasks to run as possible.

It may turn out that a --abort-on-error option is useful for development
to restore the old behaviour of aborting as soon as any errors occur ...
we'll see.